# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /connections/pending/{pendingId}:
    parameters:
      - name: pendingId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getPendingConnection
      tags: [social]
      summary: Get accounts returned by the provider that are awaiting selection
      responses:
        "200":
          description: Pending connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingConnection"
        "404":
          description: Not found or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: cancelPendingConnection
      tags: [social]
      summary: Discard a pending connection
      responses:
        "204":
          description: Discarded
  /connections/pending/{pendingId}/complete:
    parameters:
      - name: pendingId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: completePendingConnection
      tags: [social]
      summary: Connect the selected accounts from a pending connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PendingSelection"
      responses:
        "200":
          description: Accounts connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectedAccountList"
        "400":
          description: Invalid selection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /connections/{accountId}:
    parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      operationId: disconnectAccount
      tags: [social]
      summary: Disconnect an account and delete its stored tokens
      responses:
        "204":
          description: Disconnected
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /connections/{accountId}/verify:
    parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: verifyConnectedAccount
      tags: [social]
      summary: Check the account's token and permissions with the provider
      responses:
        "200":
          description: Updated account status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectedAccount"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Platform not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Platform:
      type: string
      enum: [facebook, instagram, linkedin, youtube]
    ConnectionStatus:
      type: string
      enum: [active, expired, revoked, missing_permissions, error]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
    ConnectedAccount:
      type: object
      properties:
        id:
          type: string
          format: uuid
        platform:
          $ref: "#/components/schemas/Platform"
        accountType:
          type: string
        externalAccountId:
          type: string
        displayName:
          type: string
        username:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/ConnectionStatus"
        statusDetail:
          type: ["string", "null"]
        scopes:
          type: array
          items: { type: string }
        missingScopes:
          type: array
          items: { type: string }
        firstComment:
          type: string
          enum: [supported, needs_permission, unsupported]
          description: Whether a first comment can be posted with this account.
        tokenExpiresAt:
          type: ["string", "null"]
          format: date-time
        lastVerifiedAt:
          type: ["string", "null"]
          format: date-time
        connectedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, platform, accountType, externalAccountId, displayName, username, avatarUrl, status, statusDetail, scopes, missingScopes, firstComment, tokenExpiresAt, lastVerifiedAt, connectedAt, updatedAt]
    ConnectedAccountList:
      type: object
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/ConnectedAccount"
      required: [accounts]
    PendingCandidate:
      type: object
      properties:
        externalAccountId:
          type: string
        accountType:
          type: string
        displayName:
          type: string
        username:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        selectable:
          type: boolean
        warnings:
          type: array
          items: { type: string }
        alreadyConnected:
          type: boolean
      required: [externalAccountId, accountType, displayName, username, avatarUrl, selectable, warnings, alreadyConnected]
    PendingConnection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        platform:
          $ref: "#/components/schemas/Platform"
        expiresAt:
          type: string
          format: date-time
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/PendingCandidate"
      required: [id, platform, expiresAt, candidates]
    PendingSelection:
      type: object
      properties:
        externalAccountIds:
          type: array
          minItems: 1
          maxItems: 100
          items: { type: string, minLength: 1 }
      required: [externalAccountIds]
...